+++ config/config.exs
@@ -1,13 +1,6 @@
 use Mix.Config
 
-config :guardian, Guardian.DB,
-  # Add your repository module
-  repo: GuardianRedis.Repo,
-  # default: 60 minutes
-  sweep_interval: 60
-
-# store all token types if not set
-# token_types: ["refresh_token"],
+config :guardian, Guardian.DB, adapter: GuardianRedis.Adapter
 
 config :guardian_redis, :redis,
   host: System.get_env("REDIS_HOST", "127.0.0.1"),